+++ walkability/core/operator_worker.py
@@ -86,7 +86,6 @@ class OperatorWalkability(BaseOperator[ComputeInputWalkability]):
             PointsOfInterest.SEATING: 200,
             PointsOfInterest.REMAINDER: 1000,
             PointsOfInterest.PUBLIC_TOILET: 500,
-            PointsOfInterest.SHELTERED_BENCH: 200,
         }
         self.max_walking_distance_map = {k: round(v, -1) for k, v in max_walking_distance_map.items()}
 
@@ -196,7 +195,6 @@ class OperatorWalkability(BaseOperator[ComputeInputWalkability]):
                     max_walking_distance_map=self.max_walking_distance_map,
                     ohsome_client=self.ohsome,
                     ors_settings=self.ors_settings,
-                    feature_flag_experimental=self.feature_flag_experimental,
                     resources=resources,
                 )
                 artifacts.extend(comfort_artifacts)